chore: add login required where it is necessary

+++ products/views.py
@@ -82,8 +82,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -103,7 +108,7 @@ def add_product(request):
     return render(request, template, context)
 
 
-@login_required()
+@login_required
 def add_review(request, product_id):
     """
     Renders form to add review.
@@ -166,8 +171,13 @@ def add_review(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edits a product in db """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -190,15 +200,20 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
     return redirect(reverse('products'))
 
 
-@login_required()
+@login_required
 def edit_review(request, review_id):
     """
     View to edit user reviews.
